requirements.txt
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cssselect2==0.8.0
defusedxml==0.7.1
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.61.1
fpdf2==2.8.5
idna==3.11
loguru==0.7.3
mysql-connector-python==9.5.0
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
secure-smtplib==0.1.1
six==1.17.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
urllib3==2.6.2
webencodings==0.5.1
win32_setctime==1.2.0
zopfli==0.4.0
